requirements.txt
beautifulsoup4==4.10.0
python-telegram-bot==20.3
requests==2.26.0
python-dotenv
